+++ tests/test-cran/test-record.R
@@ -24,3 +24,13 @@ assert('record() avoids recording empty plots', {
   rec = record('library(grid); g <- pointsGrob(); x <- convertUnit(g$x, "mm")')
   (length(rec) %==% 1L)  # only source code, no plot generated
 })
+
+assert('record() can selectively keep plots', {
+  f = function(i = TRUE) record(c('plot(1)', 'plot(2)'), dev.keep = i)
+  cl = function(x) sub('^record_', '', sapply(x, class))
+  (cl(f()) %==% rep(c('source', 'plot'), 2))  # keep all plots
+  (cl(f(-1)) %==% c('source', 'plot'))  # remove 1st plot
+  (cl(f(c(FALSE, TRUE))) %==% c('source', 'plot'))  # ditto
+  (cl(f(-2)) %==% c('source', 'plot', 'source'))  # remove 2nd plot
+  (cl(f(FALSE)) %==% 'source')  # remove all plots
+})