+++ NEWS.md
@@ -1,6 +1,6 @@
 # CHANGES IN xfun VERSION 0.54
 
-- `record()` no longer records empty plots triggered as side-effects by functions like `grid::convertUnit()` (thanks, @yjunechoe, yihui/litedown#96).
+- `record()` tries not to records empty plots triggered as side-effects by functions like `grid::convertUnit()` (thanks, @yjunechoe, yihui/litedown#96). However, the underlying technique is not robust enough to deal with all cases. In case certain undesired plots still get recorded, you can filter them out via the new argument of `record()`, `dev.keep`, which can take indices to determine which plots to keep. For example, `dev.keep = -1` means to remove the first plot.
 
 # CHANGES IN xfun VERSION 0.53