+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES IN xfun VERSION 0.56
+
+- Renamed the argument `url` to `input` in `browser_print()`.
+
 # CHANGES IN xfun VERSION 0.55
 
 - `record(error = TRUE)` will trim the error message by removing the uninformative part `in eval(expr, envir)` in the message (thanks, @ThomasSoeiro, yihui/litedown#109).